refactor(network): extract is_active_bft_signer + block-boundary reject
Two more cheap predicates the swarm consults before letting peer-
sourced messages into the deeper ingest pipeline. Same shape as the
multiaddr / ratelimit extractions: pulled out of the swarm event
handlers so they're easier to find + unit-test without dragging the
whole libp2p machinery along.

is_active_bft_signer is the "is this address allowed to broadcast
BFT votes / round-status?" gate. Checks DPoS first, falls back to
the legacy PoA roster.

block_boundary_reject_reason was born from the 2026-04-22 3-way
state_root fork — one validator's chain.db went bad and started
producing blocks with state_root=None above STATE_ROOT_FORK_HEIGHT.
Putting the cheap chain_id-mismatch and missing-state_root checks
at the network boundary kills those blocks before they spend a
chain write lock.

libp2p_node.rs drops further to 1860 LOC.
